Fixed TakeDamage from a mapobject
TODO: 
- Fix Combo Buff(You gain 2 combo thingy's lol)
- Fix Pets
- Fix Summon can only attack 1 mob
- Fix a lot of shet.



git-svn-id: https://svn6.assembla.com/svn/MoopleDEV@50 59296cf3-ab24-4890-a14e-c0dede3ece5d

+++ src/net/channel/handler/MovePetHandler.java
@@ -38,7 +38,7 @@ public final class MovePetHandler extends AbstractMovementPacketHandler {
             return;
         }
         MapleCharacter player = c.getPlayer();
-        int slot = player.getPetIndex(petId);
+        byte slot = player.getPetIndex(petId);
         if (slot == -1) {
             return;
         }
